requirements.txt
asgiref==3.5.2
Babel==2.11.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
cssselect2==0.7.0
Django==4.1.4
django-ckeditor==6.5.1
django-crispy-forms==2.0
django-crum==0.7.9
django-filer==2.2.4
django-jazzmin==2.6.0
django-js-asset==2.0.0
django-mptt==0.14.0
django-phonenumber-field==7.0.1
django-polymorphic==3.1.0
django-ranged-response==0.2.0
django-recaptcha==3.0.0
django-taggit==3.1.0
djangorestframework==3.14.0
easy-thumbnails==2.8.5
google-analytics-data==0.16.1
google-api-core==2.11.0
google-auth==2.16.0
googleapis-common-protos==1.58.0
GPUtil==1.4.0
grpcio==1.51.1
grpcio-status==1.51.1
gunicorn==20.1.0
idna==3.4
lxml==4.9.2
phonenumbers==8.13.3
phonenumberslite==8.13.3
Pillow==9.3.0
proto-plus==1.22.2
protobuf==4.21.12
psutil==5.9.4
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2022.7
PyYAML==6.0
reportlab==3.6.12
requests==2.28.2
rsa==4.9
six==1.16.0
sqlparse==0.4.3
svglib==1.5.1
tinycss2==1.2.1
tzdata==2022.7
Unidecode==1.1.2
urllib3==1.26.14
uuid==1.30
webencodings==0.5.1
python-dotenv==1.0.0
whitenoise==6.7.0
django-heroku==0.3.1
